backend/requirements.txt
flask
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
psycopg2-binary
